refactor: add "valid-as-of-block" in reply when getting txs to forge
When getting the top transactions from the pool with the intention to
forge them in the next block we don't want already forged transactions
to be in that list. However, the pool cannot guarantee that there will
not be forged transactions because of the nature of forging - new blocks
are forged every 8 seconds. So even if the reply from the pool is
"correct" (it does not contain forged transactions), then it could happen
that 0.1 seconds after that some of those transactions get forged.

We can guarantee that the list of returned transactions does not contain
forged ones as of some block height (H). That is - none of the returned
transactions is forged in block with height H or earlier blocks. So we
append H to the list of transactions in the reply. The forger knows the
height of the block he is forging and can decide if the reply suits him
or not. This patch changes the forger to retry a few times if the reply
is outdated and if still outdated - try to forge a block nevertheless.

This fixes https://github.com/ArkEcosystem/core/issues/1402